# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 86.88 B&B — REST API
  version: 1.0.0
  description: |
    Backend API for the 86.88 民宿 (B&B) platform.

    Serves as the bridge between the PostgreSQL database and the frontend apps:
    - `apps/website` — public-facing Next.js site (8688bnb.com)
    - `apps/admin` — admin dashboard (admin.8688bnb.com)

    ## Authentication
    Admin authentication uses JWT tokens stored in `__Host-8688_session` HttpOnly cookies.
    All state-changing requests require a CSRF token via the `X-CSRF-Token` header.

    ## Rate Limiting
    - Auth endpoints: 20 requests / 15 minutes
    - General API: 200 requests / 15 minutes
    - Public reads: 500 requests / 15 minutes
    - Webhook ingress: 100 requests / 15 minutes

    Rate limit info is returned in response headers:
    - `X-RateLimit-Limit` — max requests in window
    - `X-RateLimit-Remaining` — requests left
    - `X-RateLimit-Reset` — window reset time (Unix epoch)

  contact:
    name: 8688 BnB Dev
    email: dev@8688bnb.com
  license:
    name: Private
    identifier: UNLICENSED
servers:
  - url: http://localhost:3333/api/v1
    description: Local development
  - url: https://api.8688bnb.com/api/v1
    description: Production (via Cloudflare Tunnel)
...
security: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-8688_session
      description: |
        JWT stored in a hardened HttpOnly cookie.
        Set automatically on successful login.
        Cookie attributes: HttpOnly, Secure, SameSite=Strict, Path=/
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token required for all state-changing requests (POST, PUT, DELETE).
        Obtained from the login response or `GET /auth/csrf-token`.
    webhookApiKey:
      type: apiKey
      in: header
      name: X-Webhook-Key
      description: |
        API key for authenticating incoming OTA channel manager webhooks.
        Configured in the WEBHOOK_API_KEY environment variable.

  # ─── Reusable Schemas ────────────────────────────────────────────────────────
  schemas:
...
    AdminUser:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: yenfeng
        display_name:
          type: string
          example: 黃筵丰
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, username, display_name, created_at]
...
    BlockedDate:
      type: object
      description: |
        A date range where a room is unavailable (not a booking).
        Boss can block dates when going abroad, maintenance, etc.
        The reason is displayed as a tooltip when users hover over blocked dates.
      properties:
        id:
          type: integer
          example: 1
        room_id:
          type: integer
          nullable: true
          example: null
          description: Specific room, or null = all rooms blocked
        start_date:
          type: string
          format: date
          example: '2026-08-01'
        end_date:
          type: string
          format: date
          example: '2026-08-07'
        reason:
          type: string
          example: 老闆出國旅遊
          description: Comment/reason shown when hovering on these dates
        created_by:
          type: string
          example: yenfeng
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, start_date, end_date, reason, created_by, created_at, updated_at]
    BlockedDateCreateRequest:
      type: object
      properties:
        room_id:
          type: integer
          nullable: true
          description: Specific room ID, or omit/null to block ALL rooms
        start_date:
          type: string
          format: date
          example: '2026-08-01'
        end_date:
          type: string
          format: date
          example: '2026-08-07'
        reason:
          type: string
          minLength: 1
          maxLength: 200
          example: 老闆出國旅遊
          description: Reason displayed on hover (required so guests understand why)
      required: [start_date, end_date, reason]
    BlockedDateUpdateRequest:
      type: object
      properties:
        room_id:
          type: integer
          nullable: true
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
        reason:
          type: string
          minLength: 1
          maxLength: 200
      minProperties: 1

    # ── Media (Page-level image management) ────────────────────────────────────
...
    ApiError:
      type: object
      description: Standard error response
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
              description: |
                Machine-readable error code. Examples:
                - AUTH_INVALID_CREDENTIALS
                - AUTH_SESSION_EXPIRED
                - AUTH_CSRF_INVALID
                - VALIDATION_ERROR
                - NOT_FOUND
                - RATE_LIMITED
                - CONFLICT (e.g. double-booking)
                - INTERNAL_ERROR
                - UPLOAD_FILE_TOO_LARGE
                - UPLOAD_INVALID_TYPE
            message:
              type: string
              example: 驗證失敗
              description: Human-readable error message (Chinese)
            message_en:
              type: string
              example: Validation failed
              description: English fallback message
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: check_in
                  message:
                    type: string
                    example: 入住日期不得早於今天
              description: Per-field validation errors (if applicable)
          required: [code, message]
      required: [success, error]
    CsrfTokenResponse:
      type: object
      properties:
        csrf_token:
          type: string
          example: abc123def456...
      required: [csrf_token]
...
  parameters:
...
    BlockedDateIdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
      description: Blocked date entry ID
...
  responses:
    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: AUTH_SESSION_EXPIRED
              message: 登入已過期，請重新登入
              message_en: Session expired, please log in again
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: 找不到資源
              message_en: Resource not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: 資料驗證失敗
              message_en: Validation failed
              details:
                - field: check_in
                  message: 入住日期不得早於今天
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: RATE_LIMITED
              message: 請求太頻繁，請稍後再試
              message_en: Too many requests, please try again later
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Admin login
      description: |
        Authenticate the admin user (username: yenfeng).
        On success, sets the `__Host-8688_session` HttpOnly cookie
        and returns a CSRF token for subsequent state-changing requests.

        Rate limited: 20 attempts / 15 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: yenfeng
                password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
                  example: '********'
              required: [username, password]
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
              description: '__Host-8688_session cookie (HttpOnly, Secure, SameSite=Strict)'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/AdminUser'
                      csrf_token:
                        type: string
                        description: CSRF token — include in X-CSRF-Token header for state-changing requests
                    required: [user, csrf_token]
                required: [success, data]
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: AUTH_INVALID_CREDENTIALS
                  message: 帳號或密碼錯誤
                  message_en: Invalid username or password
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Admin logout
      description: |
        Invalidate the current session.
        Clears the `__Host-8688_session` cookie.
      security:
        - cookieAuth: []
          csrfToken: []
      responses:
        '200':
          description: Logged out successfully
          headers:
            Set-Cookie:
              schema:
                type: string
              description: Clears __Host-8688_session cookie
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 已成功登出
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Get current admin profile
      description: Returns the authenticated admin user profile.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/AdminUser'
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/password:
    put:
      tags: [Auth]
      operationId: changePassword
      summary: Change admin password
      description: |
        Change the admin password. Requires current password for verification.
        New password must meet strength requirements (min 8 chars).
        All existing sessions are invalidated after password change.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                  format: password
                  maxLength: 128
                new_password:
                  type: string
                  format: password
                  minLength: 8
                  maxLength: 128
                  description: |
                    Must be at least 8 characters.
                    All character types allowed.
              required: [current_password, new_password]
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 密碼已更新，請重新登入
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          description: Current password incorrect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: AUTH_INVALID_CREDENTIALS
                  message: 目前密碼不正確
                  message_en: Current password is incorrect
  /auth/csrf-token:
    get:
      tags: [Auth]
      operationId: getCsrfToken
      summary: Get CSRF token
      description: |
        Returns a fresh CSRF token. Useful for refreshing the token
        without re-authenticating.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: CSRF token
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/CsrfTokenResponse'
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ═══════════════════════════════════════════════════════════════════════════
  # ROOMS
  # ═══════════════════════════════════════════════════════════════════════════
...
  /blocked-dates:
    get:
      tags: [Blocked Dates]
      operationId: listBlockedDates
      summary: List blocked date ranges
      description: |
        List all blocked date ranges. Can filter by room or date range.
        Used by admin to manage room unavailability (owner travel, maintenance, etc.).
      security:
        - cookieAuth: []
      parameters:
        - name: room_id
          in: query
          schema:
            type: integer
          description: Filter by room (omit for all rooms including global blocks)
        - name: from
          in: query
          schema:
            type: string
            format: date
          description: Filter blocks overlapping this start date
        - name: to
          in: query
          schema:
            type: string
            format: date
          description: Filter blocks overlapping this end date
        - name: active_only
          in: query
          schema:
            type: boolean
            default: true
          description: Only return blocks that haven't fully passed yet
      responses:
        '200':
          description: Blocked date list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BlockedDate'
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Blocked Dates]
      operationId: createBlockedDate
      summary: Block date range
      description: |
        Block a date range for a specific room or all rooms.
        Reason is required — it's shown as a tooltip when users hover on blocked dates.
        Example reasons: "老闆出國旅遊", "房間維修", "私人使用"
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockedDateCreateRequest'
      responses:
        '201':
          description: Date range blocked
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/BlockedDate'
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Overlapping confirmed bookings exist in this range
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: CONFLICT
                  message: 此日期範圍內已有確認的預約，無法封鎖
                  message_en: Cannot block dates with existing confirmed bookings
  /blocked-dates/{id}:
    parameters:
      - $ref: '#/components/parameters/BlockedDateIdParam'

    put:
      tags: [Blocked Dates]
      operationId: updateBlockedDate
      summary: Update blocked date range
      description: Update the date range, room, or reason of a blocked period.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockedDateUpdateRequest'
      responses:
        '200':
          description: Blocked date updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/BlockedDate'
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Blocked Dates]
      operationId: deleteBlockedDate
      summary: Remove blocked date range
      description: Unblock a previously blocked date range.
      security:
        - cookieAuth: []
          csrfToken: []
      responses:
        '200':
          description: Block removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 封鎖日期已移除
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # MEDIA (Site-wide image management)
  # ═══════════════════════════════════════════════════════════════════════════
...